add router getMe

+++ Back-End/controllers/tripController.js
@@ -1,3 +1,4 @@
+const { default: mongoose } = require("mongoose");
 const Trip = require("../models/tripModel");
 
 // ----------------- get All Trips -----------------
@@ -16,8 +17,9 @@ exports.getAllTrips = async (req, res) => {
 
 //createTrip
 exports.createTrip = async (req, res) => {
+  const user_id = req.user._id;
+
   const {
-    user_id,
     driver_id,
     date,
     time,
@@ -66,6 +68,11 @@ exports.getTrip = async (req, res) => {
 
 // updateTrip
 exports.updateTrip = async (req, res) => {
+  const { id } = req.params;
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).json({ error: "No such Trip" });
+  }
   try {
     const trip = await Trip.findByIdAndUpdate(req.params.id, req.body, {
       new: true,
@@ -81,6 +88,9 @@ exports.updateTrip = async (req, res) => {
 };
 // deleteTrip
 exports.deleteTrip = async (req, res) => {
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).json({ error: "No such Trip" });
+  }
   try {
     await Trip.findByIdAndDelete(req.params.id);
     res.status(200).json({
